+++ hkl/calc.py
@@ -800,6 +800,11 @@ class CalcRecip(object):
     def __str__(self):
         return repr(self)
 
+    @property
+    def _cfg_reciprocal(self):
+        """Return reciprocal lattice to save as configuration."""
+        return tuple(list(self.sample.reciprocal))
+
 
 class CalcE4CH(CalcRecip):
     """Geometry: E4CH"""